+++ tool/travis.sh
@@ -11,5 +11,6 @@ dartanalyzer --fatal-warnings \
   lib/json_utils.dart \
   lib/version_utils.dart \
   lib/synchronized.dart \
+  lib/async_utils.dart \
 
 pub run test -p vm,firefox,chrome